+++ lib/src/domain/smac_behavior_enum.dart
@@ -1,4 +1,4 @@
-enum SmacBehavior {
+enum SmacBehaviorEnum {
   waiting,
   loading,
   success,