+++ test/mocks/smac_test_exception.dart
@@ -1,5 +1,5 @@
-import 'package:smac_dart/core/domain/smac_error.dart';
+import 'package:smac_dart/src/domain/smac_error.dart';
 
-class SmacTestException extends SmacError {
-  const SmacTestException() : super(message: 'Test error thrown');
+class SmacTestError extends SmacError {
+  const SmacTestError() : super(message: 'Test error thrown');
 }