+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 1.0.0
+* AsyncSmac changed to QuadSmac
+* QuadSmacs default waitingBuilder now displays "Unitialized"
+* GetSmacMixin now binds the Widget's BuildContext to the SMaC
+
 ## Version 0.1.0
 
 * Initial Release